Refactor PR creation script by modularizing functionality and improving code organization
