+++ utils/print_utils.py
@@ -1,0 +1,20 @@
+def red_print(text: str) -> None:
+    print(f"\033[91m{text}\033[0m")
+
+def green_print(text: str) -> None:
+    print(f"\033[92m{text}\033[0m")
+
+def yellow_print(text: str) -> None:
+    print(f"\033[93m{text}\033[0m")
+
+def blue_print(text: str) -> None:
+    print(f"\033[94m{text}\033[0m")
+
+def magenta_print(text: str) -> None:
+    print(f"\033[95m{text}\033[0m")
+
+def cyan_print(text: str) -> None:
+    print(f"\033[96m{text}\033[0m")
+
+def white_print(text: str) -> None:
+    print(f"\033[97m{text}\033[0m")
